% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{abw3eck}
\alias{abw3eck}
\title{Abwicklungsdreieck konstruieren}
\usage{
abw3eck(data, betrag, behandlungsjahr, abrechnungsjahr)
}
\arguments{
\item{data}{Datensatz mit Angaben zu den Leistungen (brutto oder netto)
sowie dem Behandlungs- und Abrechnungsjahr.}

\item{betrag}{Name der Brutto- oder Nettoleistungen in \code{data}.}

\item{behandlungsjahr}{Name des Behandlungsjahrs in \code{data}.}

\item{abrechnungsjahr}{Name des Abrechnungsjahrs in \code{data}.}
}
\value{
Matrix, die das Abwechslungsdreieck darstellt.
}
\description{
Abwicklungsdreieck konstruieren
}
