% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{additiv}
\alias{additiv}
\title{Additives Verfahren}
\usage{
additiv(S, pis)
}
\arguments{
\item{S}{S-Matrix.}

\item{pis}{Pi-Schätzungen.}
}
\value{
Mit additivem Verfahren erweiterte S-Matrix.
}
\description{
Additives Verfahren auf S-Matrix anwenden.
}
