% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{bf}
\alias{bf}
\title{Bornhuetter-Ferguson-Algorithmus}
\usage{
bf(S, gamma, alpha)
}
\arguments{
\item{S}{S-Matrix.}

\item{gamma}{Gamma-Schätzungen.}

\item{alpha}{Alpha-Schätzungen}
}
\value{
Mit Bornhuetter-Ferguson erweiterte S-Matrix.
}
\description{
Bornhuetter-Ferguson auf S-Matrix anwenden.
}
