% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{bh}
\alias{bh}
\title{Benktander-Hovinen-Algorithmus}
\usage{
bh(S, gamma, alpha)
}
\arguments{
\item{S}{S-Matrix.}

\item{gamma}{Gamma-Schätzungen.}

\item{alpha}{Alpha-Schätzungen}
}
\value{
Mit Benktander-Hovinen erweiterte S-Matrix.
}
\description{
Benktander-Hovinen auf S-Matrix anwenden.
}
