% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{cape_cod}
\alias{cape_cod}
\title{Cape cod-Verfahren}
\usage{
cape_cod(S, gamma, pis)
}
\arguments{
\item{S}{S-Matrix.}

\item{gamma}{Gamma-Schätzungen.}

\item{pis}{Pi-Schätzungen.}
}
\value{
Mit cape cod erweiterte S-Matrix.
}
\description{
Cape cod auf S-Matrix anwenden.
}
