% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{chain_ladder}
\alias{chain_ladder}
\title{Chain ladder-Algorithmus}
\usage{
chain_ladder(S)
}
\arguments{
\item{S}{S-Matrix.}
}
\value{
Mit chain ladder erweiterte S-Matrix.
}
\description{
Chain ladder auf S-Matrix anwenden.
}
