% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{chain_ladder_bf}
\alias{chain_ladder_bf}
\title{Chain ladder via Bornhuetter-Ferguson}
\usage{
chain_ladder_bf(S)
}
\arguments{
\item{S}{S-Matrix.}
}
\value{
Mit CL/BF erweiterte S-Matrix.
}
\description{
Chain ladder via Bornhuetter-Ferguson auf S-Matrix anwenden.
}
