% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eintrittsalterspraemien.R
\name{deckungskapital}
\alias{deckungskapital}
\title{Deckungskapital}
\usage{
deckungskapital(x, m, lx, i, pj, Kj)
}
\arguments{
\item{x}{Alter der Versichertenkohorte.}

\item{m}{Anzahl Jahre in die Zukunft.}

\item{lx}{Grösse der Versichertenkohorte.}

\item{i}{Zinssatz.}

\item{pj}{Vektor mit Überlebungswahrscheinlichkeiten nach Alter. Soll bei Alter 0 anfangen.}

\item{Kj}{Vektor mit erwarteten Kopfschäden nach Alter. Soll bei Alter 0 anfangen.}
}
\value{
Deckungskapital nach \code{m} Jahren.
}
\description{
Deckungskapital
}
