% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eintrittsalterspraemien.R
\name{eintrittsalterspraemie}
\alias{eintrittsalterspraemie}
\title{Eintrittsalterpraemie}
\usage{
eintrittsalterspraemie(x, i, pj, Kj)
}
\arguments{
\item{x}{Alter der Versichertenkohorte.}

\item{i}{Zinssatz.}

\item{pj}{Vektor mit Überlebungswahrscheinlichkeiten nach Alter. Soll bei Alter 0 anfangen.}

\item{Kj}{Vektor mit erwarteten Kopfschäden nach Alter. Soll bei Alter 0 anfangen.}
}
\value{
Eintrittsalterprämie.
}
\description{
Eintrittsalterpraemie
}
