% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risiko.R
\name{expected_shortfall}
\alias{expected_shortfall}
\title{Expected shortfall}
\usage{
expected_shortfall(alpha, qdist, rdist, ..., B = 1e+06)
}
\arguments{
\item{alpha}{Alpha.}

\item{qdist}{Quantilfunktion.}

\item{rdist}{Funktion, um Zufallsdaten aus Verteilung zu generieren.}

\item{...}{Parameter für \code{qdist} und \code{rdist}.}

\item{B}{Anzahl generierter Zufallsdaten.}
}
\value{
Expected shortfall.
}
\description{
Schätzt expected shortfall anhand einer Simulation.
}
