% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{im_ist_jahr}
\alias{im_ist_jahr}
\title{Abgerechnet im Ist-Jahr}
\usage{
im_ist_jahr(S, pis)
}
\arguments{
\item{S}{S-Matrix.}

\item{pis}{Volumenmass.}
}
\value{
Proportion der Kosten, die im Ist-Jahr abgerechnet werden dürften.
}
\description{
Schätzt die Proportion der Kosten, die im Ist-Jahr abgerechnet werden.
Diese Proportion wird anhand des additiven Verfahrens geschätzt.
}
