% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe.R
\name{k}
\alias{k}
\title{Individuelle Kostenbeteiligung berechnen}
\usage{
k(s, f, sb, m)
}
\arguments{
\item{s}{Leistung.}

\item{f}{Franchise.}

\item{sb}{Selbstbehalt zwischen 0 und 1.}

\item{m}{Maximaler Selbstbehalt}
}
\value{
Kostenbeteiligung.
}
\description{
Individuelle Kostenbeteiligung berechnen
}
