% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe.R
\name{k_avg}
\alias{k_avg}
\title{Kostenbeteiligung anhand von Kostenklassen berechnen.}
\usage{
k_avg(e, s, f, sb, m)
}
\arguments{
\item{e}{Vektor mit Anteilen der Kostenklassen.}

\item{s}{Vektoren mit individuellen Leistungen pro Kostenklasse.}

\item{f}{Franchise.}

\item{sb}{Selbstbehalt zwischen 0 und 1.}

\item{m}{Maximaler Selbstbehalt}
}
\value{
Kostenbeteiligung.
}
\description{
Kostenbeteiligung anhand von Kostenklassen berechnen.
}
