% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobe.R
\name{kobe_anteil}
\alias{kobe_anteil}
\title{Kostenbeteiligungsanteil anhand von Kostenklassen berechnen.}
\usage{
kobe_anteil(e, s, f, sb, m)
}
\arguments{
\item{e}{Vektor mit Anteilen der Kostenklassen.}

\item{s}{Vektoren mit individuellen Leistungen pro Kostenklasse.}

\item{f}{Franchise.}

\item{sb}{Selbstbehalt zwischen 0 und 1.}

\item{m}{Maximaler Selbstbehalt}
}
\value{
Kostenbeteiligungsanteil an Gesamtleistung.
}
\description{
Kostenbeteiligungsanteil anhand von Kostenklassen berechnen.
}
