% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eintrittsalterspraemien.R
\name{leistungsbarwert}
\alias{leistungsbarwert}
\title{Leistungsbarwert}
\usage{
leistungsbarwert(x, lx = 1, i, pj, Kj)
}
\arguments{
\item{x}{Alter der Versichertenkohorte.}

\item{lx}{Grösse der Versichertenkohorte.}

\item{i}{Zinssatz}

\item{pj}{Vektor mit Überlebungswahrscheinlichkeiten nach Alter. Soll bei Alter 0 anfangen.}

\item{Kj}{Vektor mit erwarteten Kopfschäden nach Alter. Soll bei Alter 0 anfangen.}
}
\value{
Leistungsbarwert
}
\description{
Berechnet den Leistungsbarwert.
}
