% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{loss_dev}
\alias{loss_dev}
\title{Loss development-Verfahren}
\usage{
loss_dev(S, gamma)
}
\arguments{
\item{S}{S-Matrix.}

\item{gamma}{Gamma-Schätzungen.}
}
\value{
Mit Loss development erweiterte S-Matrix.
}
\description{
Loss development-Verfahren auf S-Matrix anwenden.
}
