% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{mack}
\alias{mack}
\title{Mack-Methode}
\usage{
mack(S, pis)
}
\arguments{
\item{S}{S-Matrix.}

\item{pis}{Pi-Schätzungen}
}
\value{
Mit Mack erweiterte S-Matrix.
}
\description{
Mack auf S-Matrix anwenden.
}
