% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{panning}
\alias{panning}
\title{Panning-Verfahren}
\usage{
panning(S)
}
\arguments{
\item{S}{S-Matrix.}
}
\value{
Mit Panning erweiterte S-Matrix.
}
\description{
Panning-Verfahren auf S-Matrix anwenden.
}
