% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{phi}
\alias{phi}
\title{Phi-Faktoren berechnen}
\usage{
phi(S)
}
\arguments{
\item{S}{S-Matrix.}
}
\value{
Total Rückstellungen.
}
\description{
Anhand von S-Matrix die Phi-Faktoren berechnen.
Diese schätzen, um welchen Faktor die kumulierten
Leistungen im Vergleich zum vorigen Abrechnungsmonat
zunehmen.
}
