% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eintrittsalterspraemien.R
\name{rentenbarwert}
\alias{rentenbarwert}
\title{Rentenbarwert}
\usage{
rentenbarwert(x, lx = 1, i, pj)
}
\arguments{
\item{x}{Alter der Versichertenkohorte.}

\item{lx}{Grösse der Versichertenkohorte.}

\item{i}{Zinssatz}

\item{pj}{Vektor mit Überlebungswahrscheinlichkeiten. Soll bei Alter 0 anfangen.}
}
\value{
Rentenbarwert
}
\description{
Berechnet den Rentenbarwart.
}
