% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{rueckstellung}
\alias{rueckstellung}
\title{Gesamtrückstellungen berechnen}
\usage{
rueckstellung(S)
}
\arguments{
\item{S}{S-Matrix.}
}
\value{
Total Rückstellungen.
}
\description{
Aus S-Matrix das total der Rückstellungen berechnen.
}
