% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{shift_left}
\alias{shift_left}
\title{Matrixzeilen links aussetzen}
\usage{
shift_left(M)
}
\arguments{
\item{M}{Zu verschiebende Matrix.}
}
\value{
Matrix mit links ausgesetzten Zeilen.
}
\description{
Matrixzeilen links aussetzen
}
