% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{shift_right}
\alias{shift_right}
\title{Matrixzeilen rechts aussetzen}
\usage{
shift_right(M)
}
\arguments{
\item{M}{Zu verschiebende Matrix.}
}
\value{
Matrix mit rechts ausgesetzten Zeilen.
}
\description{
Matrixzeilen rechts aussetzen
}
