% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{to_S}
\alias{to_S}
\title{X-Matrix zu S-Matrix konvertieren.}
\usage{
to_S(X)
}
\arguments{
\item{X}{Zu konvertierende X-Matrix.}
}
\value{
Zu \code{X} passende S-Matrix. Diese Matrix ist links eingerückt.
}
\description{
X-Matrizen enthalten die Leistungen pro Zeiteinheit;
S-Matrizen enthalten die kumulierten Leistungen pro Zeiteinheit.
}
