% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{to_X}
\alias{to_X}
\title{S-Matrix zu X-Matrix konvertieren.}
\usage{
to_X(S)
}
\arguments{
\item{S}{Zu konvertierende S-Matrix.}
}
\value{
Zu \code{S} passende X-Matrix, links eingerückt.
}
\description{
X-Matrizen enthalten die Leistungen pro Zeiteinheit;
S-Matrizen enthalten die kumulierten Leistungen pro Zeiteinheit.
}
