% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rueckstellungen.R
\name{zsf}
\alias{zsf}
\title{Rückstellungszusammenfassung}
\usage{
zsf(S, alpha = NA, gamma = NA, pis)
}
\arguments{
\item{S}{S-Matrix.}

\item{alpha}{A-priori alpha-Schätzungen. Falls \code{NA} werden sie anhand des Chain Ladders geschätzt.}

\item{gamma}{A-priori gamma-Schätzungen. Falls \code{NA} werden sie anhand des Chain Ladders geschätzt.}

\item{pis}{Volumenmass.}
}
\value{
Tibble mit anhand unterschiedlicher Algorithmen geschätztem Rückstellungsbedarf.
}
\description{
Unterschiedliche Algorithmen auf S-Matrix anwenden.
Wenn \code{alpha}, \code{gamma} nicht mitgeliefert werden, werden diese anhand
des Chain-Ladders geschätzt. Falls \code{loss_dev()} mit Chain-Ladder-gamma-Schätzungen
verwendet wird, ergibt dies die Chain-Ladder-Rückstellungsschätzung.
Falls \code{bf()} mit Chain-Ladder-alpha- und -gamma-Schätzungen verwendet wird,
ergibt dies ebenfalls die Chain-Ladder-Rückstellungsschätzung.
}
